feat: added permissions management system
* added a dashboard to see and manage users permissions.
* added missing permission rules in some queries/mutations.

+++ backend/src/resolvers/Mutation.js
@@ -4,6 +4,7 @@ const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 
 const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
@@ -35,9 +36,13 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const { where } = args;
     // 1. Find the item
-    const item = await ctx.db.query.item({ where }, '{ id title }');
+    const item = await ctx.db.query.item({ where }, '{ id title user { id }}');
     // 2. Check if they own that item, or have the permissions
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+    if (!ownsItem && !hasPermissions) {
+      throw new Error('You don\'t have permission to do that!');
+    }
     // 3. Delete it
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -146,6 +151,21 @@ const Mutations = {
     });
     // 8. Return the new user
     return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // 1. Check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
+    // 2. Query the current user
+    const currentUser = await ctx.db.query.user({ where: { id: ctx.request.userId } }, info);
+    // 3. Check if they have permissions to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+    // 4. Update the permissions
+    return ctx.db.mutation.updateUser({
+      data: { permissions: { set: args.permissions } },
+      where: { id: args.userId },
+    }, info);
   },
 };
